# backend/api.yaml
openapi: 3.0.3
info:
  title: Cumulus dashboard
  description: |-
    OpenAPI specification for the Cumulus dashboard backend

    Some useful links:
    - [GitHub Repository](https://github.com/smart-on-fhir/cumulus-app)
    - [Test Instance](https://smart-cumulus-test.herokuapp.com/)

  # termsOfService: http://swagger.io/terms/
  # contact:
  #  email: apiteam@swagger.io
  license:
   name: Apache 2.0
   url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
# externalDocs:
#   description: Find out more about Swagger
#   url: http://swagger.io
servers:
  - url: https://smart-cumulus-test.herokuapp.com/api
  - url: https://smart-cumulus-ucdavis.herokuapp.com/api
  - url: https://smart-cumulus-regi-e973987fb5b9.herokuapp.com/api
  - url: https://smart-cumulus-staging.herokuapp.com/api
  - url: https://smart-cumulus.herokuapp.com/api
security: 
- session_id: []
paths:
...
  /tags:
    get:
      summary: Get all Tags
      operationId: getTags
      tags: 
        - Tags
      description: |-
        Typically used to retrieve all tags at once but some filters can be applied to restrict the returned results.

        The current user is required to have `Tags.read` permission, or or multiple explicit permissions to read every single tag record (`Tags#1.read`, `Tags#2.read`, etc.)
      parameters:
        - $ref: '#/components/parameters/where'
        - $ref: '#/components/parameters/attributes'
        - $ref: '#/components/parameters/pick'
        - $ref: '#/components/parameters/omit'
        - $ref: '#/components/parameters/include'
        - $ref: '#/components/parameters/order'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses: 
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
        '401':
          description: If the user is not authorized and guest users are not allowed to view tags
        '403':
          description: If the user is not allowed to view tags
        '400':
          description: Bad request due to invalid parameters
    post:
      summary: Creates new Tag
      operationId: createTag
      tags: 
        - Tags
      requestBody: 
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 50
                description:
                  type: string
                  minLength: 1
                  maxLength: 200
      responses:
        '200':
          description: Returns the created Tag object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: In case of invalid parameters
        '401':
          description: If the user is not authorized and guest users are not allowed to create tags
        '403':
          description: If the user is not allowed to create tags
...
  /data-sites:
    get:
      summary: Get all DataSites
      operationId: getDataSites
      tags: 
        - DataSites
      description: |-
        Typically used to retrieve all sites at once but some filters can be applied to restrict the returned results.

        The current user is required to have `DataSites.read` permission, or or multiple explicit permissions to read every single DataSites record (`DataSites#1.read`, `DataSites#2.read`, etc.)
      parameters:
        - $ref: '#/components/parameters/where'
        - $ref: '#/components/parameters/attributes'
        - $ref: '#/components/parameters/pick'
        - $ref: '#/components/parameters/omit'
        - $ref: '#/components/parameters/include'
        - $ref: '#/components/parameters/order'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses: 
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataSite'
        '401':
          description: If the user is not authorized and guest users are not allowed to view DataSites
        '403':
          description: If the user is not allowed to view DataSites
        '400':
          description: Bad request due to invalid parameters
    post:
      summary: Create new DataSite
      operationId: createDataSite
      tags: 
        - DataSites
      requestBody: 
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: '#/components/schemas/standardName'
                description:
                  $ref: '#/components/schemas/standardDescription'
      responses:
        '200':
          description: Returns the created DataSite object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSite'
        '400':
          description: In case of invalid parameters
        '401':
          description: If the user is not authorized and guest users are not allowed to create DataSites
        '403':
          description: If the user is not allowed to create DataSites
...
  /users:
    get:
      summary: Get all Users
      operationId: getUsers
      tags: 
        - Users
      description: Note that some attributes of users (like sid and password) are excluded for security reasons
      responses: 
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          description: If the client attempts to use the `where` or `include` parameters
        '401':
          description: If the user is not authorized and guest users are not allowed to view Users
        '403':
          description: If the user is not allowed to view Users
...
  /study-areas:
    get:
      summary: Get All Study Areas
      operationId: getStudyAreas
      tags: 
        - Study Areas
      description: |-
        Typically used to retrieve all study areas at once but some filters can
        be applied to restrict the returned results.

        **Note** that the response will always include the subscriptions of each
        study area and the views of each subscription.

      parameters:
        - $ref: '#/components/parameters/where'
        - $ref: '#/components/parameters/attributes'
        # - $ref: '#/components/parameters/pick'
        # - $ref: '#/components/parameters/omit'
        # - $ref: '#/components/parameters/include'
        - $ref: '#/components/parameters/order'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              example:
                id: 1
                name: Example Study Area
                description: Example Study Area Description
                creatorId: 2
                createdAt: 2024-02-22T13:36:17.315Z
                updatedAt: 2024-02-22T13:36:17.315Z
                Subscriptions: []
        '401':
          description: If the user is not authorized and guest users are not
            allowed to view included resources
        '403':
          description: If the user is not allowed to view included resources
        '400':
          description: Bad request due to invalid parameters
    post:
      summary: Create Study Area
      description: Create New Study Area
      operationId: createStudyArea
      tags: 
        - Study Areas
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                  example: View display name
                description:
                  type: string
                  example: View description
                Subscriptions:
                  type: array
                  items:
                    type: integer
                    minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              example:
                id: 1
                name: Example Study Area
                description: Example Study Area Description
                creatorId: 2
                createdAt: 2024-02-22T13:36:17.315Z
                updatedAt: 2024-02-22T13:36:17.315Z
                Subscriptions: []
        '401':
          description: If the user is not authorized and guest users are not
            allowed to create included resources
        '403':
          description: If the user is not allowed to create included resources
        '400':
          description: Bad request due to invalid parameters
        '500':
          description: Fail to create Study Area
...
  /views:
    get:
      summary: Get All Views
      description: Retrieve an array of all the views that the current user 
        is allowed to see
      operationId: getViews
      tags: 
        - Views
      parameters:
        - $ref: '#/components/parameters/order'
        - name: limit
          in: query
          schema:
            type: integer
          example: 10
          required: false
          allowEmptyValue: true
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/attributes'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/View'
                    - type: object
                      properties:
                        Tags:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                $ref: '#/components/schemas/standardID'
                              name:
                                type: string
                                maxLength: 50
                              description:
                                type: string
                                maxLength: 200
                        DataRequest:
                          type: object
                          properties:
                            id:
                              $ref: '#/components/schemas/standardID'
                            name:
                              type: string
                              maxLength: 100
              example:
                - id: 1
                  name: Example View
                  description: Example View Description
                  screenShot: Base 64 image data...
                  settings: {}
                  subscriptionId: 1
                  creatorId: 2
                  createdAt: 2024-02-22T13:36:17.315Z
                  updatedAt: 2024-02-22T13:36:17.315Z
                  Tags: []
                  DataRequest:
                    id: 1
                    name: Subscription Name
        '401':
          description: If the user is not authorized and guest users are not
            allowed to view included resources
        '403':
          description: If the user is not allowed to view included resources
        '400':
          description: Bad request due to invalid parameters
    post:
      summary: Create View
      description: Creates new View
      operationId: createView
      tags: 
        - Views
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                  example: View display name
                description:
                  type: string
                  maxLength: 500
                  example: View description
                screenShot:
                  type: string
                  example: base64imageData...
                Tags:
                  type: array
                  items:
                    $ref: '#/components/schemas/Tag'
                settings:
                  type: object
                subscriptionId:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/View'
                  - type: object
                    properties:
                      Tags:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              $ref: '#/components/schemas/standardID'
                            name:
                              type: string
                              maxLength: 50
                            description:
                              type: string
                              maxLength: 200
                      DataRequest:
                        type: object
                        properties:
                          id:
                            $ref: '#/components/schemas/standardID'
                          name:
                            type: string
                            maxLength: 100
        '401':
          description: If the user is not authorized and guest users are not
            allowed to create views
        '403':
          description: If the user is not allowed to create views
        '400':
          description: Bad request due to invalid parameters
    delete:
      summary: Delete Multiple Views
      operationId: bulkDeleteView
      description: Deletes multiple Views
      tags: 
        - Views
      parameters:
        - name: id
          in: query
          required: true
          description: Comma-separated list of IDs to delete
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              example:
                message: Deleted 5 graphs
        '401':
          description: If the user is not authorized and guest users are not
            allowed to delete included resources
        '403':
          description: If the user is not allowed to delete included resources
        '400':
          description: Bad request due to invalid parameters
...
  /requests:
    get:
      summary: Get All Subscriptions 
      operationId: getAllSubscriptions
      tags:
        - Subscriptions
      parameters:
        - $ref: '#/components/parameters/where'
        - $ref: '#/components/parameters/attributes'
        - $ref: '#/components/parameters/pick'
        - $ref: '#/components/parameters/omit'
        - $ref: '#/components/parameters/include'
        - $ref: '#/components/parameters/order'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'
        '401':
          description: If the user is not authorized and guest users are not
            allowed to view included resources
        '403':
          description: If the user is not allowed to view included resources
    post:
      summary: Create new Subscription
      description: Creates a Subscription. If an array of Tags is provided,
        attaches them to the created subscription.
      operationId: createSubscription
      tags:
        - Subscriptions
      requestBody: 
        content:
          application/json:
            example: {
              "name": "string",
              "description": "string",
              "refresh": "manually",
              "groupId": 1,
              "dataURL": "string",
              # "transmissions": {}
            }
      responses:
        '200':
          description: Returns the created Subscription object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: In case of invalid parameters
        '401':
          description: If the user is not authorized and guest users are not allowed to create Subscriptions
        '403':
          description: If the user is not allowed to create Subscriptions
        '500':
          description: If we fail to create the Subscription
...
  /request-groups:
    get:
      summary: Get all Subscription Groups
      operationId: getSubscriptionGroups
      tags: 
        - Subscription Groups
      parameters:
        - name: subscriptions
          in: query
          description: If set, includes the Subscriptions with their id, name, completed, and the name of every View in that subscription
          schema:
            type: boolean
      responses: 
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionGroup'
        '401':
          description: If the user is not authorized and guest users are not allowed to view SubscriptionGroups
        '403':
          description: If the user is not allowed to view SubscriptionGroups
        '400':
          description: Bad request due to invalid parameters
    post:
      summary: Creates new Subscription Group
      operationId: createSubscriptionGroup
      tags: 
        - Subscription Groups
      requestBody: 
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                description:
                  type: string
      responses:
        '200':
          description: Returns the created SubscriptionGroup object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionGroup'
        '400':
          description: In case of invalid parameters
        '401':
          description: If the user is not authorized and guest users are not allowed to create SubscriptionGroups
        '403':
          description: If the user is not allowed to create SubscriptionGroups
...
  /permissions:
    get:
      summary: Get All Permissions
      operationId: getAllPermissions
      tags:
        - Permissions
      parameters:
        - $ref: '#/components/parameters/where'
        - $ref: '#/components/parameters/attributes'
        - $ref: '#/components/parameters/pick'
        - $ref: '#/components/parameters/omit'
        - $ref: '#/components/parameters/include'
        - $ref: '#/components/parameters/order'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses: 
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'
        '401':
          description: If the user is not authorized and guest users are not allowed to view Permissions
        '403':
          description: If the user is not allowed to view Permissions
        '400':
          description: Bad request due to invalid parameters
    delete:
      summary: Bulk Delete
      operationId: deleteMultiplePermissions
      tags:
        - Permissions
      parameters:
        - name: id
          in: query
          required: true
          description: Comma-separated list of IDs to delete
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              example:
                message: Deleted 5 records
        '401':
          description: If the user is not authorized and guest users are not
            allowed to delete included resources
        '403':
          description: If the user is not allowed to delete included resources
        '400':
          description: Bad request due to invalid parameters

  # /permissions/revoke:
  #   post:
  #     summary: Revoke Permission
  #     description: Removes permission(s) matched by the input parameters
  #     operationId: revokePermission
  #     tags:
  #       - Permissions
  #     requestBody:
  #       content:
  #         application/json:
  #           schema:
  #             type: object
  #             properties:
  #               resource:
  #                 description: The resource type
  #                 type: string
  #               resource_id:
  #                 description: If set, it means we are managing only the
  #                   resource(s) with the specified ID(s). Can be a number or an
  #                   array of numbers or null. Defaults to null, which means all
  #                   resources. If used, all the IDs must be found in DB.
  #                 oneOf:
  #                   - type: number
  #                     nullable: true
  #                   - type: array
  #                     nullable: true
  #                     items:
  #                       type: number
  #               email:
  #                 description: The email(s) of the user(s) we are going to modify.
  #                   Can be a string or an array of strings or null. If set,
  #                   then user_group_id and role must be null or omitted! **This
  #                   cannot be used together with `role` or `user_group_id`!**
  #                   If used, all emails must be found in DB.
  #                 oneOf:
  #                   - type: string
  #                     nullable: true
  #                   - type: array
  #                     nullable: true
  #                     items:
  #                       type: string
  #               role:
  #                 type: array
  #                 nullable: true
  #                 items:
  #                   type: string
  #                 description: The role(s) we are going to share with. Can be an
  #                   array of strings or null. If set, then email and
  #                   user_group_id must be null or omitted! **This cannot be used
  #                   together with `email` or `user_group_id`!**
  #               user_group_id:
  #                 description: The ID(s) of the userGroup(s) we are going to
  #                   share with. Can be a number or an array of numbers or null.
  #                   If set, then email and role must be null or omitted! **This
  #                   cannot be used together with `email` or `role`!** If used,
  #                   all the IDs must be found in DB.
  #                 oneOf:
  #                 - type: number
  #                   nullable: true
  #                 - type: array
  #                   nullable: true
  #                   items:
  #                       type: number
  #     responses:
  #       '200':
  #         description: Successful operation
  #         content:
  #           application/json:
  #             example: { ok: true }
  #       '400':
  #         description: In case of bad parameters
  #       '403':
  #         description: You are not allowed to perform this action
  #       '500':
  #         description: Failed operation
  #         content:
  #           application/json:
  #             example: { ok: false }
...
  /user-groups:
    get:
      summary: Get all User Groups
      operationId: getUserGroups
      tags: 
        - User Groups
      description: Get all User Groups
      parameters:
        - $ref: '#/components/parameters/where'
        - $ref: '#/components/parameters/attributes'
        - $ref: '#/components/parameters/pick'
        - $ref: '#/components/parameters/omit'
        # - $ref: '#/components/parameters/include'
        - $ref: '#/components/parameters/order'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/UserGroup'
                    - type: object
                      properties:
                        users:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                $ref: '#/components/schemas/standardID'
                              name:
                                type: string
                                maxLength: 100
                              email:
                                type: string
                                format: email
                              role:
                                type: string
                                enum:
                                  - user
                                  - manager
                                  - admin
        '401':
          description: If the user is not authorized and guest users are not
            allowed to view included resources
        '403':
          description: If the user is not allowed to view included resources
    post:
      summary: Create new User Group
      description: Create an user group with the given name and description. If
        an array of users is provided, sets them as the user whi belong to this
        group.
      operationId: createUserGroup
      tags: 
        - User Groups
      requestBody: 
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  $ref: '#/components/schemas/standardName'
                description:
                  $ref: '#/components/schemas/standardDescription'
                users:
                  type: array
                  items:
                    $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Returns the created DataSite object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSite'
        '400':
          description: In case of invalid parameters
        '401':
          description: If the user is not authorized and guest users are not allowed to create User Groups
        '403':
          description: If the user is not allowed to create User Groups
        '500':
          description: If we fail to create the User Group
...
components:
  parameters:
    where:
      name: where
      in: query
      schema:
        type: string
      description: |-
        Can contain zero or more comma-separated where conditions.
        
        #### *Examples*:
        - ***name:value** pairs - `name:John`*
        - ***name:operator:value** - `age:lt:10`*
        - ***multiple conditions** - `name:John,age:gt:5`*
      required: false
      allowEmptyValue: true
    attributes:
      name: attributes
      in: query
      schema:
        type: string
      description: |-
        Comma-separated list of model attributes to include in the response

        ***Example**: `id, name, age`*
      required: false
      allowEmptyValue: true
    pick:
      name: pick
      in: query
      schema:
        type: string
      description: |-
        Comma-separated list of model attributes to include in the response.
        
        **NOTE:** This parameter is ignored when the `attributes` parameter is used

        ***Example**: `id, name, age`*
      required: false
      allowEmptyValue: true
    omit:
      name: omit
      in: query
      schema:
        type: string
      description: |-
        Comma-separated list of model attributes to exclude from the response.
        
        **NOTE:** This parameter is ignored when the `attributes` parameter is used

        ***Example**: `id, name, age`*
      required: false
      allowEmptyValue: true
    include:
      name: include
      in: query
      schema:
        type: string
      description: |-
        Comma-separated list of associations to include. Each item in the list can be:
        
        ***Examples**:*
        - *`model` - The name of the model to include*
        - *`model:attribute1|attribute2...` - Specify what attributes to include for each model*
        - *`model:attribute1:alias1|attribute2...` - Can also specify aliases for included attributes*
      required: false
      allowEmptyValue: true
    order:
      name: order
      in: query
      schema: 
        type: string
      description: |-
        Sorting instruction(s)

        ***Example**: `name:ask,age:desc`*
      required: false
      allowEmptyValue: true
    limit:
      name: limit
      in: query
      schema:
        type: string
      description: |-
        limit or limits for each included model

        ***Examples**:*
        - Basic limit: `10`
        - Also specify limits for included models: `10,Users:30`
      required: false
      allowEmptyValue: true
    offset:
      name: offset
      in: query
      description: How many records to skip
      required: false
      allowEmptyValue: true
      schema: 
        type: integer
  schemas:
...
    Tag:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/standardID'
        name:
          $ref: '#/components/schemas/standardName'
        description:
          $ref: '#/components/schemas/standardDescription'
        creatorId:
          $ref: '#/components/schemas/standardID'
        createdAt:
          $ref: '#/components/schemas/autoTimestamp'
        updatedAt:
          $ref: '#/components/schemas/autoTimestamp'
...
  securitySchemes:
    session_id:
      type: apiKey
      name: sid
      in: cookie